feat: implement real-time hourly sales tracking and date filtering for dashboard

+++ src/main/java/com/fusioncrew/aikiosk/domain/admin/dto/DashboardSummaryDto.java
@@ -15,7 +15,7 @@ public class DashboardSummaryDto {
     private Double revenueChange;
     private Integer activeOrders;
     private List<LowStockIngredientDto> lowStockIngredients;
-    private List<Integer> hourlySales;
+    private List<Long> hourlySales;
     private List<ActivityLogDto> aiActivityLogs;
 
     @Data
